PastelComCoca - New target selector
Provides methods to Lock an especific champion for example:
You are fighting 2v2 on the botlane and the enemy marksman is one hit
and runs into the bush... instead of hitting the support wasting an
important auto attack this class can lock your focus on the adc, so when
you get vision you will have the auto attack up.
Just double click the champion you want to lock.

Preview
http://i.imgur.com/d1gk2MI.png
http://i.imgur.com/2RJeEdH.png
http://i.imgur.com/GAmteqU.png
http://i.imgur.com/VWlJcwi.png
http://i.imgur.com/cbNIyO2.png
http://i.imgur.com/ZjajoWz.png